- set clean restrictions after update from websocket and import - add icon url back when displaying
